license-server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-multipart==0.0.20
jinja2==3.1.4
python-dotenv==1.0.1
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
httpx==0.28.1

# Production DB (Render provides DATABASE_URL for Postgres)
psycopg2-binary==2.9.10
